refactor: folder structure + eslint rules for imports
