+++ package.json
@@ -35,12 +35,15 @@
         "@types/sinon-chrome": "^2.2.15",
         "@types/webpack": "^5.28.5",
         "@types/webpack-env": "^1.18.8",
+        "@typescript-eslint/parser": "^8.25.0",
         "clean-webpack-plugin": "^4.0.0",
         "copy-webpack-plugin": "^12.0.2",
         "css-loader": "^7.1.2",
         "eslint": "^9.19.0",
         "eslint-config-prettier": "^10.0.1",
+        "eslint-import-resolver-typescript": "^3.8.3",
         "eslint-plugin-check-file": "^3.0.0",
+        "eslint-plugin-import": "^2.31.0",
         "eslint-plugin-prettier": "^5.2.3",
         "eslint-plugin-react": "^7.37.4",
         "html-webpack-plugin": "^5.6.3",
